[PATCH] USB Storage: fix Genesys Logic based on info from vendor
In theory, this is the fix we need to make Genesys Logic devices work.
This patch started life as as343, which was created based on some
information which a user finally coaxed out of Genesys Logic.  Limited
end-user testing gives good results.

As we expected, it's a bug in their device.  This is really a workaround
for what is almost certainly a timing problem.  Apparently, the 'popular'
OSes don't push the device as hard as Linux does.

Although it is likely that this workaround is not needed for all Genesys
devices, Genesys was unable/unwilling to provide us with the explicit list
of VID/PIDs which required this -- thus we apply it to all Genesys devices.

We have lots of good reports with max_sectors set to 128 with these
devices, but the official recommendation is to set that to 64.  End-users
can experiment with higher values (for higher performance) via the runtime
sysfs interface to that parameter.

I would like to give special thanks to the users who hounded Genesys into
giving up this information, and to Alan Stern for not giving up on this
vendor long after I had.

Signed-off-by: Alan Stern <stern@rowland.harvard.edu>
Signed-off-by: Matthew Dharm <mdharm-usb@one-eyed-alien.net>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>