[PATCH] sparc32: gcc-3.3 macro parenthesization fix for memcpy.S
From: Art Haas <ahaas@airmail.net>

The 1.3->1.4 changes to the arch/sparc/lib/copy_user.S file added parenthesis
to a number of macros within that file.  The BK changlog associated with this
change indicate the change was to make the file work with gcc-3.3.

When looking at the changes made, I see that similar macros exist in memcpy.S
as well, so would a patch adding parens to that file be worthwhile?  Also,
just what was the problem with gcc-3.3 that was resolved by adding the
parenthesis?  Macro mis-expansion I'm guessing.

Signed-off-by: Art Haas <ahaas@airmail.net>
Signed-off-by: William Irwin <wli@holomorphy.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>