kbuild: __crc_* symbols in System.map
David S. Miller <davem@redhat.com> wrote:
Shouldn't we be grepping __crc_ symbols out of the System.map file?

For one thing, these can confuse readprofile.  It's algorithm is
to start at _stext, then stop when it sees a line in the System.map
which is not text (mode is one of 'T' 't' 'W' or 'w')

It will exit early if there are some intermixed __crc_* things in
there (since they are are mode 'A').

For example, in my current sparc64 kernel I have this:

00000000004cef80 t do_split
00000000004cf2a0 t add_dirent_to_buf
00000000004cf5a7 A __crc_init_special_inode
00000000004cf640 t make_indexed_dir
00000000004cf900 t ext3_add_entry

So no symbols after add_dirent_to_buf will be shown in the profiling
output of readprofile.

Implementation ported to mksysmap by Sam.
Included two System.map related fixes:
- Print "SYSMAP  System.map" during build
- Sort symbols in System.map

Signed-off-by: Sam Ravnborg <sam@ravnborg.org>