[PATCH] tmscsim: (CH) Fix error handling
I might be wrong, but it looks like the attached patch has still not been
applied, although it is pretty important - it fixes error handling in
tmscsim. E.g., it was reported, that this patch fixes a problem with
tmscsim in 2.6.7 with a SCSI-3 scanner, that doesn't support REPORT LUNS
command.

Signed-off-by: James Bottomley <James.Bottomley@SteelEye.com>