[PATCH] ppc64: set time-related systemcfg fields
Somewhere along the line we lost the code that updates some fields of the
systemcfg structure that are used for translating timebase values to time
of day.  I want to get rid of the systemcfg structure eventually, but
applications are using it (and in particular these fields) and I don't want
to break the ABI in a stable kernel series.

Signed-off-by: Paul Mackerras <paulus@samba.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>