[PATCH] RCU - cpu-offline-cleanup
There is a series of patches in my tree and these 3 are the first ones that
should probably be merged down the road.  Descriptions are on top of the
patches.  Please include them in -mm.

A lot of RCU code will be cleaned up later in order to support
call_rcu_bh(), the separate RCU interface that considers softirq handler
completion a quiescent state.

This patch:

Minor cleanup of the hotplug code to remove #ifdef in cpu event notifier
handler.  If CONFIG_HOTPLUG_CPU is not defined, CPU_DEAD case will be
optimized off.

Signed-off-by: Dipankar Sarma <dipankar@in.ibm.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>